database/requirements.txt
fastapi
uvicorn
google-cloud-firestore
pydantic